refactor: move SpringSinkRow into distinct mod
